updated entry_cards to cards
